Use the transactional endpoint.
With the 'resultDataContents' parameter, and the 'graph' contents it enables,
we can now use the transactional endpoint for all of the needs of the browser.

Since this format includes the labels of nodes, we can skip the filter concept,
and don't have to take the extra roundtrip for retrieving the labels.